[Refactor] : Tuist Plugin 추가
ProjectEnvironment 생성

Git : #2
