+++ App/iOS/Project.swift
@@ -1,24 +1,5 @@
-//
-//  Project.swift
-//  ProjectDescriptionHelpers
-//
-//  Created by minii on 2023/05/23.
-//
-
 import ProjectDescription
-import ProjectDescriptionHelpers
-
-//let project = Project.makeModule(
-//  name: "Sample",
-//  platform: .iOS,
-//  product: .app,
-//  dependencies: [
-//    .project(target: "Feature", path: .relativeToRoot("Projects/Feature"))
-//  ],
-//  resources: ["Resources/**"],
-//  infoPlist: .file(path: "Support/Info.plist")
-//)
-
+import ProjectEnvironment
 
 let project = Project.makeModule(
   name: "iOS",
