+++ Plugins/ProjectEnvironment/Plugin.swift
@@ -1,0 +1,3 @@
+import ProjectDescription
+
+let plugin = Plugin(name: "ProjectEnvironment")
