+++ Tuist/Config.swift
@@ -1,7 +1,7 @@
 import ProjectDescription
 
 let config = Config(
-//    plugins: [
-//        .local(path: .relativeToManifest("../../Plugins/AnimatedDrawings")),
-//    ]
+    plugins: [
+        .local(path: .relativeToManifest("../../Plugins/ProjectEnvironment"))
+    ]
 )
